move pin clearance helper to utils
